fix(analytics): keep measures total consistent with the sum of its facet buckets
(cherry picked from commit 79128f150af9117d6f2a28d496efdaf022c82c18)

+++ gravitee-apim-repository/gravitee-apim-repository-elasticsearch/src/test/java/io/gravitee/repository/elasticsearch/tracing/ElasticsearchTracingRepositoryTest.java
@@ -375,7 +375,7 @@ class ElasticsearchTracingRepositoryTest {
         assertThat(page.getTotalElements()).isEqualTo(200L);
     }
 
-    private static Aggregation cardinalityAggregation(float value) {
+    private static Aggregation cardinalityAggregation(double value) {
         Aggregation agg = new Aggregation();
         agg.setValue(value);
         return agg;
